refactor(web): break circular dependency between api and apiWithRetry
Extract API_BASE into lib/apiConfig.ts and the CSRF token cache and
helpers into lib/csrf.ts, so apiWithRetry.ts can import refreshCsrfToken
statically instead of doing a lazy import("./api") to dodge the cycle.

api.ts re-exports API_BASE, getCsrfToken and refreshCsrfToken, so the
~20 modules importing them from @/lib/api need no changes.
